Refactor MenuLayout
