+++ src/layouts/MenuLayout/components/index.js
@@ -1,3 +1,4 @@
 // @flow strict
 
+export { Back } from './Back/Back'
 export { MenuPanel } from './MenuPanel/MenuPanel'
